+++ packages/vant-cli/src/common/types.ts
@@ -1,5 +1,0 @@
-import type Webpack from 'webpack';
-
-export type WebpackConfig = Webpack.Configuration & {
-  devServer?: any;
-};